countedNodes changed to double - has really great values

+++ datatypes.h
@@ -51,7 +51,7 @@ typedef struct cgpProperties {
 	int cols;			//count of columns in the matrix
 	int l_back; 		//l-back param of the matrix
 	int individCount;	//count of individuals in the generation
-	long countedNodes;
+	double countedNodes;
 }TCgpProperties;
 
 //whole CGP struct
